logger test files moved to lib/logger
